ability upload profile pic, user can only delete their own profile

+++ hello.py
@@ -12,6 +12,9 @@ from flask_sqlalchemy import SQLAlchemy
 from datetime import datetime, date
 from flask_migrate import Migrate
 from werkzeug.security import generate_password_hash, check_password_hash
+from werkzeug.utils import secure_filename
+import uuid as uuid
+import os 
 from flask_login import (
     UserMixin, 
     login_user, 
@@ -49,6 +52,10 @@ app.config['SQLALCHEMY_DATABASE_URI'] = 'mysql+pymysql://root:password123@localh
 app.app_context().push()
     #Create a Secret Key
 app.config['SECRET_KEY'] = "SuperDuper-Secret-Key"
+
+    #Upload Profile Pictures to Images Folder
+UPLOAD_FOLDER = 'static/images/'
+app.config['UPLOAD_FOLDER'] = UPLOAD_FOLDER
 
     #Initialize the Database
 db = SQLAlchemy(app)
@@ -129,24 +136,29 @@ def dashboard():
 @app.route('/delete/<int:id>')
 @login_required
 def delete(id):
-    user_to_delete = Users.query.get_or_404(id)
-    name = None
-    form = UserForm()
-    our_users = Users.query.all()
-    try:
-        db.session.delete(user_to_delete)
-        db.session.commit()
-        flash("Success! User was Deleted")
-        return redirect(url_for('add_user'))
-        
-    except:
-        flash("Whoops! There was a problem deleting user. Please try again!")
-        return render_template(
-        "add_user.html",
-        form=form,
-        name=name,
-        our_users=our_users
-        )
+    if id == current_user.id:
+        user_to_delete = Users.query.get_or_404(id)
+        name = None
+        form = UserForm()
+        our_users = Users.query.all()
+        try:
+            db.session.delete(user_to_delete)
+            db.session.commit()
+            flash("Success! User was Deleted")
+            return redirect(url_for('add_user'))
+            
+        except:
+            flash("Whoops! There was a problem deleting user. Please try again!")
+            return render_template(
+            "add_user.html",
+            form=form,
+            name=name,
+            our_users=our_users
+            )
+    else:
+        flash("Sorry, You Can Only Delete Your Profile")
+        return redirect(url_for('dashboard'))
+
 
 
     #Create Login Route
@@ -294,8 +306,21 @@ def update(id):
         name_to_update.email = request.form['email']
         name_to_update.favorite_color = request.form['favorite_color']
         name_to_update.about_author = request.form['about_author']
+        name_to_update.profile_pic = request.files['profile_pic']
+
+            #Grab The Image Name
+        pic_filename = secure_filename(name_to_update.profile_pic.filename)
+            #Set UUID
+        pic_name = str(uuid.uuid1()) + "_" + pic_filename
+            #Save The Image
+        saver = request.files['profile_pic']
+        
+            #Change to String to Save to Database
+        name_to_update.profile_pic = pic_name
+
         try: #Try to Commit to Database
             db.session.commit()
+            saver.save(os.path.join(app.config['UPLOAD_FOLDER']), pic_name)
             flash("User Updated!")
             return redirect(url_for('add_user'))
             #return render_template("update.html",
@@ -453,6 +478,7 @@ class Users(db.Model, UserMixin):
     favorite_color = db.Column(db.String(120))
     about_author = db.Column(db.Text(500), nullable=True)
     date_added = db.Column(db.DateTime, default=datetime.utcnow)
+    profile_pic = db.Column(db.String(300), nullable=True)
         #Password Stuff!
     password_hash = db.Column(db.String(128))
     #Users Can Have Many Posts
